tagging package snapd version 2.35.2

+++ tests/lib/reset.sh
@@ -4,20 +4,43 @@ set -e -x
 
 # shellcheck source=tests/lib/dirs.sh
 . "$TESTSLIB/dirs.sh"
+# shellcheck source=tests/lib/state.sh
+. "$TESTSLIB/state.sh"
+
+
+# shellcheck source=tests/lib/systemd.sh
+. "$TESTSLIB/systemd.sh"
+
+#shellcheck source=tests/lib/systems.sh
+. "$TESTSLIB"/systems.sh
 
 reset_classic() {
     # Reload all service units as in some situations the unit might
     # have changed on the disk.
     systemctl daemon-reload
 
-    systemctl stop snapd.service snapd.socket
+    echo "Ensure the service is active before stopping it"
+    retries=20
+    systemctl status snapd.service snapd.socket || true
+    while systemctl status snapd.service snapd.socket | grep "Active: activating"; do
+        if [ $retries -eq 0 ]; then
+            echo "snapd service or socket not active"
+            exit 1
+        fi
+        retries=$(( retries - 1 ))
+        sleep 1
+    done
+
+    systemd_stop_units snapd.service snapd.socket
 
     case "$SPREAD_SYSTEM" in
         ubuntu-*|debian-*)
             sh -x "${SPREAD_PATH}/debian/snapd.postrm" purge
             ;;
-        fedora-*|opensuse-*)
-            sh -x "${SPREAD_PATH}/packaging/fedora/snap-mgmt.sh" \
+        fedora-*|opensuse-*|arch-*|amazon-*)
+            # We don't know if snap-mgmt was built, so call the *.in file
+            # directly and pass arguments that will override the placeholders
+            sh -x "${SPREAD_PATH}/cmd/snap-mgmt/snap-mgmt.sh.in" \
                 --snap-mount-dir="$SNAP_MOUNT_DIR" \
                 --purge
             # The script above doesn't remove the snapd directory as this
@@ -45,19 +68,18 @@ reset_classic() {
     rm -f /tmp/core* /tmp/ubuntu-core*
 
     if [ "$1" = "--reuse-core" ]; then
-        # Purge all the systemd service units config
-        rm -rf /etc/systemd/system/snapd.service.d
-        rm -rf /etc/systemd/system/snapd.socket.d
-
         # Restore snapd state and start systemd service units
-        tar -C/ -xzf "$SPREAD_PATH/snapd-state.tar.gz"
+        restore_snapd_state
         escaped_snap_mount_dir="$(systemd-escape --path "$SNAP_MOUNT_DIR")"
-        mounts="$(systemctl list-unit-files --full | grep "^$escaped_snap_mount_dir[-.].*\.mount" | cut -f1 -d ' ')"
-        services="$(systemctl list-unit-files --full | grep "^$escaped_snap_mount_dir[-.].*\.service" | cut -f1 -d ' ')"
+        mounts="$(systemctl list-unit-files --full | grep "^${escaped_snap_mount_dir}[-.].*\\.mount" | cut -f1 -d ' ')"
+        services="$(systemctl list-unit-files --full | grep "^${escaped_snap_mount_dir}[-.].*\\.service" | cut -f1 -d ' ')"
         systemctl daemon-reload # Workaround for http://paste.ubuntu.com/17735820/
         for unit in $mounts $services; do
             systemctl start "$unit"
         done
+
+        # force all profiles to be re-generated
+        rm -f /var/lib/snapd/system-key
     fi
 
     if [ "$1" != "--keep-stopped" ]; then
@@ -65,10 +87,10 @@ reset_classic() {
 
         # wait for snapd listening
         EXTRA_NC_ARGS="-q 1"
-        if [[ "$SPREAD_SYSTEM" = fedora-* ]]; then
+        if [[ "$SPREAD_SYSTEM" = fedora-* || "$SPREAD_SYSTEM" = amazon-* ]]; then
             EXTRA_NC_ARGS=""
         fi
-        while ! printf "GET / HTTP/1.0\r\n\r\n" | nc -U $EXTRA_NC_ARGS /run/snapd.socket; do sleep 0.5; done
+        while ! printf 'GET / HTTP/1.0\r\n\r\n' | nc -U $EXTRA_NC_ARGS /run/snapd.socket; do sleep 0.5; done
     fi
 }
 
@@ -77,28 +99,43 @@ reset_all_snap() {
     # shellcheck source=tests/lib/names.sh
     . "$TESTSLIB/names.sh"
 
+    remove_bases=""
+    # remove all app snaps first
     for snap in "$SNAP_MOUNT_DIR"/*; do
         snap="${snap:6}"
         case "$snap" in
-            "bin" | "$gadget_name" | "$kernel_name" | core | README)
+            "bin" | "$gadget_name" | "$kernel_name" | "$core_name" | README)
                 ;;
             *)
-                snap remove "$snap"
+                # make sure snapd is running before we attempt to remove snaps, in case a test stopped it
+                if ! systemctl status snapd.service snapd.socket; then
+                    systemctl start snapd.service snapd.socket
+                fi
+                if ! echo "$SKIP_REMOVE_SNAPS" | grep -w "$snap"; then
+                    if snap info "$snap" | egrep '^type: +(base|core)'; then
+                        remove_bases="$remove_bases $snap"
+                    else
+                        snap remove "$snap"
+                    fi
+                fi
                 ;;
         esac
     done
+    # remove all base/os snaps at the end
+    if [ -n "$remove_bases" ]; then
+        snap remove "$remove_bases"
+    fi
 
     # ensure we have the same state as initially
     systemctl stop snapd.service snapd.socket
-    rm -rf /var/lib/snapd/*
-    tar -C/ -xzf "$SPREAD_PATH/snapd-state.tar.gz"
+    restore_snapd_state
     rm -rf /root/.snap
     if [ "$1" != "--keep-stopped" ]; then
         systemctl start snapd.service snapd.socket
     fi
 }
 
-if [[ "$SPREAD_SYSTEM" == ubuntu-core-16-* ]]; then
+if is_core_system; then
     reset_all_snap "$@"
 else
     reset_classic "$@"
@@ -106,6 +143,6 @@ fi
 
 if [ "$REMOTE_STORE" = staging ] && [ "$1" = "--store" ]; then
     # shellcheck source=tests/lib/store.sh
-    . $TESTSLIB/store.sh
+    . "$TESTSLIB"/store.sh
     teardown_staging_store
 fi
